Contact.java: Extracted common fields/methods.
Common fields and methods were extracted from Address, Email and Phone.java. Replaced all usage of direct reference to the field(s) via its respective getter.
With this, code duplication is reduced!
