+++ test/java/seedu/addressbook/parser/ParserTest.java
@@ -285,9 +285,9 @@ public class ParserTest {
     private static String convertPersonToAddCommandString(ReadOnlyPerson person) {
         String addCommand = "add "
                 + person.getName().fullName
-                + (person.getPhone().isPrivate() ? " pp/" : " p/") + person.getPhone().value
-                + (person.getEmail().isPrivate() ? " pe/" : " e/") + person.getEmail().value
-                + (person.getAddress().isPrivate() ? " pa/" : " a/") + person.getAddress().value;
+                + (person.getPhone().isPrivate() ? " pp/" : " p/") + person.getPhone().getValue()
+                + (person.getEmail().isPrivate() ? " pe/" : " e/") + person.getEmail().getValue()
+                + (person.getAddress().isPrivate() ? " pa/" : " a/") + person.getAddress().getValue();
         for (Tag tag : person.getTags()) {
             addCommand += " t/" + tag.tagName;
         }
